[refactor] folder for each sql type
