+++ sql-blind/blind_sql.py
